+++ docs/rules/__tests__/attribute-boolean-style.md.spec.ts
@@ -1,0 +1,19 @@
+import HtmlValidate from '../../../src/htmlvalidate';
+
+const markup: {[key: string]: string} = {};
+markup["incorrect"] = `<input required="">
+<input required="required">`;
+markup["correct"] = `<input required>`;
+
+describe('docs/rules/attribute-boolean-style.md', () => {
+	it('inline validation: incorrect', () => {
+		const htmlvalidate = new HtmlValidate({"rules":{"attribute-boolean-style":"error"}});
+		const report = htmlvalidate.validateString(markup["incorrect"]);
+		expect(report.results).toMatchSnapshot();
+	});
+	it('inline validation: correct', () => {
+		const htmlvalidate = new HtmlValidate({"rules":{"attribute-boolean-style":"error"}});
+		const report = htmlvalidate.validateString(markup["correct"]);
+		expect(report.results).toMatchSnapshot();
+	});
+});